Quit but ugly fix to restore SLAM.
WordCreation need the shared_ptr of the self dataframe, but we do not have it and neither can create a new shared_ptr (it destroyed the real ptr when destroying the second instance of the shared_ptr). So... an easy solution is to receive the original shared_ptr. It works as expected but obviously, it is not very elegant.